规范

+++ src/main/java/com/example/vnollxonlinejudge/judge/JudgeStrategy.java
@@ -1,7 +1,7 @@
 package com.example.vnollxonlinejudge.judge;
 
-import com.example.vnollxonlinejudge.common.result.RunResult;
+import com.example.vnollxonlinejudge.model.result.RunResult;
 
 public interface JudgeStrategy {
-    RunResult judge(String code, String dataZipUrl, int timeLimit, int memoryLimit);
+    RunResult judge(String code, String dataZipUrl, Long timeLimit, Long memoryLimit);
 }
