+++ src/main/java/com/example/vnollxonlinejudge/service/JudgeService.java
@@ -1,6 +1,10 @@
 package com.example.vnollxonlinejudge.service;
-import com.example.vnollxonlinejudge.common.result.Result;
 public interface JudgeService {
 
-    String judgeSubmit(String code, String option, long pid, long uid, long cid, String create_time, String uname, int time, int memory);
+    String judgeSubmit(
+            String code, String option,
+            Long pid, Long uid, Long cid,
+            String create_time, String uname,
+            Long time, Long memory
+    );
 }
